move theme selector to backpex hooks
