+++ assets/js/hooks/index.js
@@ -1,3 +1,4 @@
 export { default as BackpexCancelEntry } from './_cancel_entry'
 export { default as BackpexDragHover } from './_drag_hover'
 export { default as BackpexSidebarSections } from './_sidebar_sections'
+export { default as BackpexThemeSelector } from './_theme_selector'
